[Fix] #33  내 그룹 상세 화면 그룹 내 랭킹 화면 수정

+++ app/src/main/java/com/team/score/API/response/group/GroupRankingResponse.kt
@@ -14,6 +14,6 @@ data class RankerInfo(
     val rankNum: Int,
     val changedAmount: Int,
     val weeklyLevelIncrement: Int,
-    val weeklyExerciseTime: Int
+    val weeklyExerciseTime: Double
 )
